WASM gating. Split shared macros into common file
