+++ zarrs/Cargo.toml
@@ -36,7 +36,6 @@ async = ["dep:async-generic", "dep:async-trait", "dep:futures", "moka/future", "
 jiff = ["dep:jiff"] # Adds support for jiff time array elements
 chrono = ["dep:chrono"] # Adds support for chrono time array elements
 float8 = ["dep:float8"] # Adds support for float8 array elements
-wasm = ["uuid", "quick_cache"]
 
 [lints]
 workspace = true
@@ -64,7 +63,6 @@ inventory.workspace = true
 itertools = "0.14.0"
 lru = "0.16.0"
 moka = { version = "0.12.8", features = ["sync"] }
-quick_cache = { version = "0.6.16", optional = true }
 ndarray = { version = ">=0.15.4,<17", optional = true }
 num.workspace = true
 pco = { version = "0.4.0", optional = true }
@@ -92,10 +90,13 @@ chrono = { version = "0.4.39", optional = true }
 float8 = { version = "0.4.1", optional = true, features = ["bytemuck"] }
 simd-adler32 = { version = "0.3.7", optional = true }
 itoa = "1.0.15"
-uuid = { version = "^1.1.0", features = ["v4", "js"], optional = true }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+quick_cache = { version = "0.6.16" }
+uuid = { version = "^1.1.0", features = ["v4", "js"] }
 
 # Reference: https://github.com/kylebarron/parquet-wasm/blob/ad18b841a8e788b0d4b17eb9deaca581d02bf887/Cargo.toml#L100C1-L110C18
-[dependencies.getrandom_v03]
+[target.'cfg(target_arch = "wasm32")'.dependencies.getrandom_v03]
 package = "getrandom"
 version = "^0.3.1"
 features = ["wasm_js"]
