+++ zarrs/src/array/chunk_cache/chunk_cache_lru_moka.rs
@@ -15,11 +15,15 @@ use crate::{
     array::{
         codec::ArrayToBytesCodecTraits, Array, ArrayBytes, ArrayError, ArrayIndices, ArraySize,
         ChunkCacheTypePartialDecoder,
-        chunk_cache::{ChunkCache, ChunkCacheType, ChunkCacheTypeDecoded, ChunkCacheTypeEncoded}
+        chunk_cache::{ChunkCache, ChunkCacheType, ChunkCacheTypeDecoded, ChunkCacheTypeEncoded},
     },
     array_subset::ArraySubset,
     storage::StorageError,
+    impl_ChunkCacheLruEncoded,
+    impl_ChunkCacheLruDecoded,
+    impl_ChunkCacheLruPartialDecoder,
 };
+
 
 use std::borrow::Cow;
 
@@ -285,7 +289,7 @@ macro_rules! impl_ChunkCacheLruCommon {
         }
     };
 }
-
+/*
 macro_rules! impl_ChunkCacheLruEncoded {
     () => {
         fn retrieve_chunk(
@@ -372,6 +376,7 @@ macro_rules! impl_ChunkCacheLruEncoded {
         }
     };
 }
+    */
 
 impl ChunkCache for ChunkCacheEncodedLruChunkLimit {
     impl_ChunkCacheLruEncoded!();
@@ -383,6 +388,7 @@ impl ChunkCache for ChunkCacheEncodedLruSizeLimit {
     impl_ChunkCacheLruCommon!();
 }
 
+/*
 macro_rules! impl_ChunkCacheLruDecoded {
     () => {
         fn retrieve_chunk(
@@ -437,6 +443,7 @@ macro_rules! impl_ChunkCacheLruDecoded {
         }
     };
 }
+    */
 
 impl ChunkCache for ChunkCacheDecodedLruChunkLimit {
     impl_ChunkCacheLruDecoded!();
@@ -446,55 +453,6 @@ impl ChunkCache for ChunkCacheDecodedLruChunkLimit {
 impl ChunkCache for ChunkCacheDecodedLruSizeLimit {
     impl_ChunkCacheLruDecoded!();
     impl_ChunkCacheLruCommon!();
-}
-
-macro_rules! impl_ChunkCacheLruPartialDecoder {
-    () => {
-        fn retrieve_chunk(
-            &self,
-            chunk_indices: &[u64],
-            options: &crate::array::codec::CodecOptions,
-        ) -> Result<ChunkCacheTypeDecoded, ArrayError> {
-            let partial_decoder = self
-                .try_get_or_insert_with(chunk_indices.to_vec(), || {
-                    self.array.partial_decoder(chunk_indices)
-                })
-                .map_err(|err| {
-                    // moka returns an Arc'd error, unwrap it noting that ArrayError is not cloneable
-                    Arc::try_unwrap(err).unwrap_or_else(|err| {
-                        ArrayError::StorageError(StorageError::from(err.to_string()))
-                    })
-                })?;
-            let chunk_shape =
-                crate::array::chunk_shape_to_array_shape(&self.array.chunk_shape(chunk_indices)?);
-            Ok(partial_decoder
-                .partial_decode(&ArraySubset::new_with_shape(chunk_shape), options)?
-                .into_owned()
-                .into())
-        }
-
-        fn retrieve_chunk_subset(
-            &self,
-            chunk_indices: &[u64],
-            chunk_subset: &ArraySubset,
-            options: &crate::array::codec::CodecOptions,
-        ) -> Result<ChunkCacheTypeDecoded, ArrayError> {
-            let partial_decoder = self
-                .try_get_or_insert_with(chunk_indices.to_vec(), || {
-                    self.array.partial_decoder(chunk_indices)
-                })
-                .map_err(|err| {
-                    // moka returns an Arc'd error, unwrap it noting that ArrayError is not cloneable
-                    Arc::try_unwrap(err).unwrap_or_else(|err| {
-                        ArrayError::StorageError(StorageError::from(err.to_string()))
-                    })
-                })?;
-            Ok(partial_decoder
-                .partial_decode(chunk_subset, options)?
-                .into_owned()
-                .into())
-        }
-    };
 }
 
 impl ChunkCache for ChunkCachePartialDecoderLruChunkLimit {
